# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: COIMAGEN OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workflow-templates:
    get:
      operationId: listWorkflowTemplates
      tags: [workflow-templates]
      summary: List all workflow templates
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowTemplate"
    post:
      operationId: createWorkflowTemplate
      tags: [workflow-templates]
      summary: Create a workflow template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowTemplateCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowTemplate"
  /workflow-templates/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      operationId: getWorkflowTemplate
      tags: [workflow-templates]
      summary: Get template by id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowTemplate"
        "404":
          description: Not found
    patch:
      operationId: updateWorkflowTemplate
      tags: [workflow-templates]
      summary: Update a workflow template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowTemplateUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowTemplate"
    delete:
      operationId: deleteWorkflowTemplate
      tags: [workflow-templates]
      summary: Delete a template
      responses:
        "204":
          description: Deleted

  # ─── TASKS ─────────────────────────────────────────────────────────────────
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      parameters:
        - { name: projectId, in: query, required: false, schema: { type: integer } }
        - { name: agentId, in: query, required: false, schema: { type: integer } }
        - { name: status, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    patch:
      operationId: updateTask
      tags: [tasks]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    delete:
      operationId: deleteTask
      tags: [tasks]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── DASHBOARD ─────────────────────────────────────────────────────────────
...
  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      parameters:
        - { name: clientId, in: query, required: false, schema: { type: integer } }
        - { name: status, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}:
    get:
      operationId: getSubscription
      tags: [subscriptions]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "404":
          description: Not found
    patch:
      operationId: updateSubscription
      tags: [subscriptions]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "404":
          description: Not found
    delete:
      operationId: deleteSubscription
      tags: [subscriptions]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── REVENUE ───────────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    Task:
      type: object
      required: [id, title, status, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        projectId: { type: ["integer", "null"] }
        projectName: { type: ["string", "null"] }
        agentId: { type: ["integer", "null"] }
        agentName: { type: ["string", "null"] }
        status: { type: string, enum: [todo, in_progress, review, done, cancelled] }
        priority: { type: string, enum: [low, medium, high, urgent] }
        dueDate: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: ["string", "null"] }
    TaskInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        projectId: { type: integer }
        agentId: { type: integer }
        status: { type: string, enum: [todo, in_progress, review, done, cancelled], default: todo }
        priority: { type: string, enum: [low, medium, high, urgent], default: medium }
        dueDate: { type: string }
    TaskUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        projectId: { type: integer }
        agentId: { type: integer }
        status: { type: string, enum: [todo, in_progress, review, done, cancelled] }
        priority: { type: string, enum: [low, medium, high, urgent] }
        dueDate: { type: string }

    # ── DASHBOARD ─────────────────────────────────────────────────────────────
...
    Subscription:
      type: object
      required: [id, plan, amount, billingCycle, status, createdAt]
      properties:
        id: { type: integer }
        clientId: { type: ["integer", "null"] }
        clientName: { type: ["string", "null"] }
        plan: { type: string }
        amount: { type: number }
        billingCycle: { type: string, enum: [monthly, quarterly, annual] }
        status: { type: string, enum: [active, paused, cancelled] }
        startDate: { type: ["string", "null"] }
        nextBillingDate: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: ["string", "null"] }
    SubscriptionInput:
      type: object
      required: [plan, amount]
      properties:
        clientId: { type: integer }
        plan: { type: string, minLength: 1 }
        amount: { type: number }
        billingCycle: { type: string, enum: [monthly, quarterly, annual], default: monthly }
        status: { type: string, enum: [active, paused, cancelled], default: active }
        startDate: { type: string }
        nextBillingDate: { type: string }
        notes: { type: string }
    SubscriptionUpdate:
      type: object
      properties:
        clientId: { type: integer }
        plan: { type: string, minLength: 1 }
        amount: { type: number }
        billingCycle: { type: string, enum: [monthly, quarterly, annual] }
        status: { type: string, enum: [active, paused, cancelled] }
        startDate: { type: string }
        nextBillingDate: { type: string }
        notes: { type: string }

    # ── REVENUE ───────────────────────────────────────────────────────────────
...
    WorkflowTemplate:
      type: object
      required: [id, name, stages, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string, nullable: true }
        product: { type: string, nullable: true }
        stages:
          type: array
          items: { type: string }
        defaultPriority: { type: string }
        isDefault: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string, nullable: true }
    WorkflowTemplateCreate:
      type: object
      required: [name, stages]
      properties:
        name: { type: string }
        description: { type: string }
        product: { type: string }
        stages:
          type: array
          items: { type: string }
        defaultPriority: { type: string }
        isDefault: { type: boolean }
    WorkflowTemplateUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        product: { type: string }
        stages:
          type: array
          items: { type: string }
        defaultPriority: { type: string }
        isDefault: { type: boolean }
...